Add files via upload
From the comment text of the automation (duration: approx. 21 min.):

Each of the above modules contains three pairs of sine wave oscillators that begin at the module's indicated target frequency. The primary pair of oscillators of a given module symmetrically detune as per the indicated beating range factor (range 0-3 in this automation). 

The secondary pair of oscillators take the frequency of one of the primary oscillators as their target frequency and also symmetrically deviate from it as per the indicated beating range factor. The third pair of oscillators recapitulate this relationship with reference to the frequency of one of the secondary oscillators. In effect, the beating range factor determines the harmonic range of an oscillator pair-set of an above module. 

A sine function dynamically scales the rate of oscillator frequency change. 

Because beating range is indicated as a proportion of target frequency, rates of frequency beats among oscillator pairs are not necessarily synchronized. 

In this automation, the beating range factor increases with time as does its rate of increase. This increases the harmonic range of the output. 

At intervals, beating range increase implies increase of the target frequency of the second module via multiplication of the top module's target frequency with an incremented value (range 4-18 in this automation). This also increases the harmonic range of the output.

At intervals of decreasing duration, the change of oscillator frequencies pauses for a duration and then resumes. Such a pause provides a sort of snapshot of transition, to paraphrase Bergson, with such snapshots themselves dynamic in their periodicity. The length of the pause duration decreases as time elapses. 

--
This patch embeds:
https://github.com/bblohowiak/Max/blob/main/2021-12-27CounterSine.maxpat
https://github.com/bblohowiak/Max/blob/main/2021-12-28DynamicFreqBeater.maxpat
https://github.com/bblohowiak/Max/blob/main/2021-12-28TripleDFB.maxpat